Refactor like the wind
