+++ src/main/scala/fi/vm/sade/hakurekisteri/integration/koski/KoskiService.scala
@@ -210,7 +210,7 @@ class KoskiService(virkailijaRestClient: VirkailijaRestClient,
     val filteredHenkilot = removeOpiskeluoikeudesWithoutDefinedOppilaitosAndOppilaitosOids(henkilot)
     if(filteredHenkilot.nonEmpty) {
       Future.sequence(filteredHenkilot.map(henkilo =>
-        koskiDataHandler.processHenkilonKoskiSuoritukset(henkilo, personOidsWithAliases.intersect(henkilo.henkilö.oid.toSet), params).recoverWith {
+        koskiDataHandler.processHenkilonTiedotKoskesta(henkilo, personOidsWithAliases.intersect(henkilo.henkilö.oid.toSet), params).recoverWith {
           case e: Exception =>
             logger.error("Koskisuoritusten tallennus henkilölle {} epäonnistui: {} ",henkilo.henkilö.oid , e)
             Future.failed(e)
