+++ src/test/scala/fi/vm/sade/hakurekisteri/integration/koski/KoskiActorSpec.scala
@@ -41,6 +41,7 @@ class KoskiActorSpec extends FlatSpec with Matchers with FutureWaiting with Spec
     }
   })
   val koskiDataHandler: KoskiDataHandler = new KoskiDataHandler(testRef, testRef, testRef)
+  val opiskelijaParser = new KoskiOpiskelijaParser
   val params: KoskiSuoritusHakuParams = new KoskiSuoritusHakuParams(true, false)
 
   //todo make sure disabling this is ok. We are now assuming that some optionals are actually always present.
@@ -52,21 +53,21 @@ class KoskiActorSpec extends FlatSpec with Matchers with FutureWaiting with Spec
   }*/
 
   it should "detectOppilaitos should return 10 as luokka for peruskoulun lisäopetus" in {
-    koskiDataHandler.detectOppilaitos(
+    opiskelijaParser.detectOppilaitos(
       SuoritusLuokka(VirallinenSuoritus(Oids.lisaopetusKomoOid, "orgId", "VALMIS", parseLocalDate("2017-01-01"), "henkilo_oid",
         yksilollistaminen.Ei, "FI", None, true, OrganisaatioOids.oph, None, Map.empty), "", parseLocalDate("2017-01-01"))
     ) should equal ("10", "orgId", "10")
   }
 
   it should "detectOppilaitos should return luokka for peruskoulun lisäopetus if not empty" in {
-    koskiDataHandler.detectOppilaitos(
+    opiskelijaParser.detectOppilaitos(
       SuoritusLuokka(VirallinenSuoritus(Oids.lisaopetusKomoOid, "orgId", "VALMIS", parseLocalDate("2017-01-01"), "henkilo_oid",
         yksilollistaminen.Ei, "FI", None, true, OrganisaatioOids.oph, None, Map.empty), "10C", parseLocalDate("2017-01-01"))
     ) should equal ("10", "orgId", "10C")
   }
 
   it should "createOpiskelija should create opiskelija" in {
-    koskiDataHandler.createOpiskelija("henkilo_oid",
+    opiskelijaParser.createOpiskelija("henkilo_oid",
       SuoritusLuokka(VirallinenSuoritus(Oids.perusopetusKomoOid, "orgId", "VALMIS", parseLocalDate("2017-01-01"), "henkilo_oid",
         yksilollistaminen.Ei, "FI", None, true, OrganisaatioOids.oph, None, Map.empty), "9F", parseLocalDate("2016-01-01"), Some("9"))
       ) should equal (
